move working_time_counter.h
